+++ src/features/Feed/components/CommentList/ChildCommentItem.tsx
@@ -6,7 +6,7 @@ import { useProfile } from '@/hooks/useProfile';
 import useCommentStore from '@/store/commentStore';
 import LikeButton from '@/components/Button/LikeButton';
 import { motion } from 'framer-motion';
-import { formatDate } from '@/utils/format';
+import { formatDate } from '@/utils/date';
 import ProfileHeader from '../ProfileHeader';
 import { Comments } from '../../types/comment';
 import { useChangeChildComment } from '../../hooks/child/useChangeChildComment';
