+++ src/features/Profile/components/Profile/ProfileEditForm.tsx
@@ -1,14 +1,14 @@
 'use client';
 
+import React from 'react';
 import TitleHeader from '@/components/Header/TitleHeader';
 import MainLayout from '@/layouts/MainLayout';
 import ProfileForm from '@/components/Profile/ProfileForm';
-import { getToday } from '@/features/Join/utils/date';
-import React from 'react';
 import { FormProvider, useForm } from 'react-hook-form';
 import ImageEditor from './ImageEditor';
 import { useProfileEdit } from '../../hooks/useProfileEdit';
 import { ProfileEditFormType } from '../../types/editForm';
+import { getToday } from '@/utils/date';
 
 export const defaultValue: ProfileEditFormType = {
   username: '',
